backend/pyproject.toml
[project]
name = "calendarai-backend"
version = "0.1.0"
description = "AI Calendar App Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-ai-slim[openrouter]>=0.1",
    "pydantic-settings>=2.0",
    "supabase>=2.0",
    "python-jose[cryptography]>=3.3",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "icalendar>=6.0",
    "python-dateutil>=2.9",
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "semantic-router>=0.1",
    "sentence-transformers>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
